Configure `auto-release` by default (#2498)
* Allow DEV_ARTIFACT_ID to write custom artifact id

* Prevent removing package mock

* Create app for auto-release

+++ packages/create-yoshi-app/src/sentry-registration/addSentry.ts
@@ -5,12 +5,16 @@ import Listr from 'listr';
 // We need to mock pkg.json in current directory, so add-sentry tool will find this file and use for configuration
 const PKG_FILE_NAME = path.join(process.cwd(), 'package.json');
 const addPackageJSONMock = () => {
+  if (fs.existsSync(PKG_FILE_NAME)) {
+    return false;
+  }
   const tempPkgContent = JSON.stringify({
     devDependencies: {
       typescript: 'latest',
     },
   });
   fs.writeFileSync(PKG_FILE_NAME, tempPkgContent);
+  return true;
 };
 
 const removePackageJSONMock = () => {
@@ -24,14 +28,16 @@ export const addSentry = ({
   teamName: string;
   projectName: string;
 }) => {
-  addPackageJSONMock();
+  const packageMockWasAdded = addPackageJSONMock();
   const tasks = require('@wix/add-sentry/lib/tasks');
   const {
     configureSentry,
     handleError,
   } = require('@wix/add-sentry/lib/sentry');
 
-  removePackageJSONMock();
+  if (packageMockWasAdded) {
+    removePackageJSONMock();
+  }
   configureSentry();
 
   const tasksToExecute = [
